whisper_finetuning/pyproject.toml
[tool.poetry]
name = "whisper-finetuning"
version = "0.1.0"
description = "Finetuning Whisper speech model for German"
authors = ["Ali Saghiran <a.saghiran@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
transformers = {extras = ["torch"], version = "^4.42.4"}
datasets = "^2.20.0"
pytorch-lightning = "^2.3.3"
jiwer = "^3.0.4"
soundfile = "^0.12.1"
librosa = "^0.10.2.post1"
tensorboardx = "^2.6.2.2"
tensorboard = "^2.17.0"
ipykernel = "^6.29.5"
evaluate = "^0.4.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
nbconvert = "^7.16.4"

[tool.black]
line-length = 79

[tool.flake8]
max-line-length = 79
exclude = [
    # No need to traverse our git directory
    ".git",
    ".venv",
    # There's no value in checking cache directories
    "__pycache__"
]
# Use extend-ignore to add to already ignored checks which are anti-patterns like W503.
extend-ignore = [
    # PEP 8 recommends to treat : in slices as a binary operator with the lowest priority, and to leave an equal
    # amount of space on either side, except if a parameter is omitted (e.g. ham[1 + 1 :]).
    # This behaviour may raise E203 whitespace before ':' warnings in style guide enforcement tools like Flake8.
    # Since E203 is not PEP 8 compliant, we tell Flake8 to ignore this warning.
    # https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices    
    "E203"
]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
